⚡️ [성능 향상] : Get API 호출에 대한 Redis 캐싱 퍼포먼스 약식 Check = 응답속도 94.8% 감소(575ms->20~30ms)
