+++ modules/bootstrap/api-payment-gateway/src/main/kotlin/im/bigs/pg/api/payment/PaymentController.kt
@@ -31,8 +31,6 @@ class PaymentController(
     private val paymentUseCase: PaymentUseCase,
     private val queryPaymentsUseCase: QueryPaymentsUseCase,
 ) {
-    /** 결제 생성 요청 페이로드(간소화된 필드). */
-    /** API 응답을 위한 변환용 DTO. 도메인 모델을 그대로 노출하지 않습니다. */
 
     /**
      * 결제 생성.
